+++ loader.php
@@ -1,7 +1,6 @@
 <?php
 
-/**
- * Add applicable namespaces to the ProcessWire classLoader.
- */
-wire('classLoader')->addNamespace('Rockett', __DIR__ . '/src/Rockett');
+namespace ProcessWire;
+
+wire('classLoader')->addNamespace('Util', __DIR__ . '/src/Util');
 wire('classLoader')->addNamespace('Typographer', __DIR__ . '/src/Typographer');