+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0
+
+- Upgrade `mundschenk-at/php-typography` to `^6.7`
+- Require PHP >= 7.4 and ProcessWire >= 3.0
+- Add Composer support (`rockett/textformatter-typographer`)
+- Reduced doc blocks and added type-hints and return types
+
 ## 1.1.0
 
 - Added `formatString` method to use the formatter outside of the Field context.